Fix withdraw from null cursorTarget exploit
We need to pass the original cursorTarget object reference from fn_selfActions.sqf to this script. 

When there is some client lag there is a small delay between the initial cursorTarget reading in fn_selfActions.sqf and the new one in this file. That allows a player to quickly look away from the original storage object and end up with a null cursorTarget. The result is a "can not get vehicle capacity" error message, but the bank dialog stays up. The player could then withdraw any amount.